made repository not directly linked to doctrine because it is possible to have documents directly from external feed.
Removed the WrappedSearchDocumentRepositoryInterface because this could
be done with a custom repository.

Changed the SearchDocumentRepositoryAdapter to EntityRepositoryWrapper
and act as bridge between an doctrine repository and solr repository.

Changed SearchDocumentRepositoryInterface to DocumentRepositoryInterface
and because of the doctrine dependency drop the where argument is
removed and when needed this should be handled by an custom repository
or just implement BaseQueryBuilderRepositoryInterface